Move roster modification stuff to its own file
