+++ commands/assign.py
@@ -2,7 +2,7 @@ import logging
 import tempfile
 import time
 
-from assigner import get_filtered_roster
+from roster_util import get_filtered_roster
 from baserepo import BaseRepo, StudentRepo
 from config import config_context
