+++ commands/open.py
@@ -2,7 +2,7 @@ import logging
 
 from requests.exceptions import HTTPError
 
-from assigner import get_filtered_roster
+from roster_util import get_filtered_roster
 from baserepo import Access, Repo, RepoError, StudentRepo
 from config import config_context
